authentication

+++ controllers/feed.js
@@ -2,6 +2,8 @@ const fs = require('fs')
 const path = require('path')
 
 const Post = require('../models/post')
+const User = require('../models/user')
+
 const { validationResult } = require('express-validator')
 
 exports.getPosts = (req, res, next) => {
@@ -39,15 +41,31 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path
     const title = req.body.title
     const content = req.body.content
+    let creator
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'Omkar' }
+        creator: req.userId
     })
     post.save()
     .then(result => {
-        res.status(200).json({message: 'Post created successfully', post: result})
+        return User.findById(req.userId)
+    })
+    .then(user => {
+        creator = user
+        user.posts.push(post)
+        return user.save()
+    })
+    .then(result => {
+        res.status(200).json({
+            message: 'Post created successfully', 
+            post: post, 
+            creator: {
+                _id: creator._id,
+                name: creator.name
+            }
+        })
     })
     .catch(err => {
         if (!err.statusCode) {
@@ -102,6 +120,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404
             throw error
         }
+        if (post.creator.toString() !== req.userId.toString()) {
+            const error = new Error('Not Authorized.')
+            error.statusCode = 403
+            throw error
+        }
         if(imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl)
         }
@@ -140,8 +163,20 @@ exports.deletePost = (req, res, next) => {
             error.statusCode = 404
             throw error
         }
+        if (post.creator.toString() !== req.userId ){
+            const error = new Error('Not Authorized.')
+            error.statusCode = 403
+            throw error
+        }
         clearImage(post.imageUrl)
         return Post.findByIdAndRemove(postId)
+    })
+    .then(result => {
+        return User.findById(req.userId)
+    })
+    .then(user => {
+        user.posts.pull(postId)
+        return user.save()
     })
     .then(result => {
         console.log(result)
